+++ C/binder/Binder_Specifiers.cpp
@@ -23,8 +23,8 @@
 #include "SyntaxTree.h"
 
 #include "binder/Scope.h"
-#include "binder/Semantics_TypeQualifiers.h"
-#include "binder/Semantics_TypeSpecifiers.h"
+#include "binder/SemanticsOfTypeQualifiers.h"
+#include "binder/ConstraintsInTypeSpecifiers.h"
 #include "compilation/SemanticModel.h"
 #include "symbols/Symbol_ALL.h"
 #include "symbols/SymbolName_ALL.h"
