+++ C/binder/ConstraintsInDeclarators.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "Semantics_Declarators.h"
+#include "ConstraintsInDeclarators.h"
 
 using namespace psy;
 using namespace C;
