+++ C/binder/ConstraintsInTypeSpecifiers.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "Semantics_TypeSpecifiers.h"
+#include "ConstraintsInTypeSpecifiers.h"
 
 #include "SyntaxTree.h"
